+++ app/src/main/java/woowacourse/movie/data/dummy/MovieData.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.data.dummy
 
 import androidx.annotation.DrawableRes
 import woowacourse.movie.R
