+++ app/src/main/java/woowacourse/movie/data/dummy/TheaterData.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.data.dummy
 
 import woowacourse.movie.domain.model.Theater
