+++ app/src/main/java/woowacourse/movie/data/repository/MovieRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.data.repository
 
-import woowacourse.movie.data.MovieData
+import woowacourse.movie.data.dummy.MovieData
 import woowacourse.movie.domain.model.Movie
 
 interface MovieRepository {
