+++ app/src/main/java/woowacourse/movie/data/ticket/TicketEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data.entity
+package woowacourse.movie.data.ticket
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
