+++ app/src/main/java/woowacourse/movie/presentation/common/ItemClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.common.adapter
+package woowacourse.movie.presentation.common
 
 fun interface ItemClickListener<T> {
     fun onClick(item: T)
