+++ app/src/main/java/woowacourse/movie/util/StringFormatter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.common.util
+package woowacourse.movie.util
 
 import java.time.LocalDate
 import java.time.LocalDateTime
