+++ app/src/test/java/woowacourse/movie/fixture/repository/TicketData.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.fixture.repository
 
-import woowacourse.movie.data.MovieData
-import woowacourse.movie.data.TheaterData
+import woowacourse.movie.data.dummy.MovieData
+import woowacourse.movie.data.dummy.TheaterData
 import woowacourse.movie.domain.model.HeadCount
 import woowacourse.movie.domain.model.Ticket
 import woowacourse.movie.domain.model.seat.Seat
